Add support for custom headers.
According to the Developer Guide it is possible to set the following
non-metadata headers for an object:
 * Content-Disposition
 * Content-Encoding
 * Access-Control-Allow-Credentials
 * Access-Control-Allow-Methods
 * Access-Control-Allow-Origin
 * Access-Control-Expose-Headers
 * Access-Control-Max-Age
 * Access-Control-Request-Headers
 * Access-Control-Request-Method
 * Origin

This patch adds an extra $headers member to CF_Object and the necessary
code to set these headers on the objects.  I didn't add any magic to the
$metadata field to avoid breaking backwards compatibility.

There is one small change in behaviour:  Adding the same (metadata) header
twice overrides the previous value and isn't ignored anymore.  The
chances of this happening are almost nil though and the checks for
duplicate headers weren't complete in the first place.  This code might
also throw a few more exceptions where it just broke down silently
before.

I didn't run the unit tests (yet) since I failed to set up PHPUnit.  The
copy and move code paths are thus untested, the unit tests aren't
updated.